+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Renamed `PeerHandle::next_message()` to `recv_message()`.
 - Moved message body out of `ReceivedRequest`.
 - Renamed `Incoming` to `ReceivedMessage`.
+- Removed unused `Outgoing` type.
 
 ## v0.4.2 - 2021-05-20
 ### Fixed